add a list parameter for subtractions
