+++ changelog_entry.yaml
@@ -2,3 +2,5 @@
   changes:
     added:
     - Oklahoma military retirement benefit exclusion.
+    - Oklahoma AGI subtractions list. 
+
