+++ policyengine_us/variables/gov/states/ok/tax/income/ok_agi_subtractions.py
@@ -8,17 +8,9 @@ class ok_agi_subtractions(Variable):
     unit = USD
     definition_period = YEAR
     reference = (
-        "https://oklahoma.gov/content/dam/ok/en/tax/documents/forms/individuals/past-year/2021/511-Pkt-2021.pdf"
-        "https://oklahoma.gov/content/dam/ok/en/tax/documents/forms/individuals/current/511-Pkt.pdf"
+        "https://oklahoma.gov/content/dam/ok/en/tax/documents/forms/individuals/past-year/2021/511-Pkt-2021.pdf#page=16"
+        "https://oklahoma.gov/content/dam/ok/en/tax/documents/forms/individuals/current/511-Pkt.pdf#page=16"
     )
     defined_for = StateCode.OK
 
-    def formula(tax_unit, period, parameters):
-        oasdi = add(tax_unit, period, ["taxable_social_security"])
-        pensions = tax_unit.members("taxable_pension_income", period)
-        p = parameters(period).gov.states.ok.tax.income.agi.subtractions
-        capped_pension = tax_unit.sum(min_(p.pension_limit, pensions))
-        military_retirement_exclusion = tax_unit(
-            "ok_military_retirement_exclusion", period
-        )
-        return oasdi + capped_pension + military_retirement_exclusion
+    adds = "gov.states.ok.tax.income.agi.subtractions.subtractions"
